refactor: 도메인 분리 - member, security
