--- springboot-monolithic/src/main/java/springbootmonolithic/domain/member/command/application/service/AuthService.java
@@ -1,7 +1,0 @@
-package springbootmonolithic.domain.member.command.application.service;
-
-import springbootmonolithic.domain.member.command.application.dto.request.SignUpRequestDTO;
-
-public interface AuthService {
-    void signUp(SignUpRequestDTO signupRequestDTO);
-}
