+++ springboot-monolithic/src/main/java/springbootmonolithic/domain/member/command/application/service/MemberService.java
@@ -1,0 +1,7 @@
+package springbootmonolithic.domain.member.command.application.service;
+
+import springbootmonolithic.domain.member.command.application.dto.request.SignUpRequestDTO;
+
+public interface MemberService {
+    void signUp(SignUpRequestDTO signupRequestDTO);
+}
