+++ springboot-monolithic/src/main/java/springbootmonolithic/domain/member/command/application/service/MemberServiceImpl.java
@@ -1,0 +1,56 @@
+package springbootmonolithic.domain.member.command.application.service;
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
+import org.springframework.stereotype.Service;
+import springbootmonolithic.common.domain.member.role.RoleType;
+import springbootmonolithic.domain.member.command.application.dto.request.SignUpRequestDTO;
+import springbootmonolithic.domain.member.command.domain.aggregate.entity.Member;
+import springbootmonolithic.domain.member.command.domain.aggregate.entity.MemberRole;
+import springbootmonolithic.domain.member.command.domain.repository.MemberRepository;
+import springbootmonolithic.domain.member.command.domain.repository.MemberRoleRepository;
+import springbootmonolithic.domain.member.query.service.MemberQueryService;
+
+@Service
+public class MemberServiceImpl implements MemberService {
+
+    private final MemberRepository memberRepository;
+    private final MemberRoleRepository memberRoleRepository;
+    private final MemberQueryService memberQueryService;
+    private final BCryptPasswordEncoder passwordEncoder;
+
+    @Autowired
+    public MemberServiceImpl(
+            MemberRepository memberRepository,
+            MemberRoleRepository memberRoleRepository,
+            MemberQueryService memberQueryService,
+            BCryptPasswordEncoder passwordEncoder
+    ) {
+        this.memberRepository = memberRepository;
+        this.memberRoleRepository = memberRoleRepository;
+        this.memberQueryService = memberQueryService;
+        this.passwordEncoder = passwordEncoder;
+    }
+
+    @Override
+    public void signUp(SignUpRequestDTO signupRequestDTO) {
+        memberQueryService.validateEmail(signupRequestDTO.getEmail());
+        memberQueryService.checkNicknameDuplicated(signupRequestDTO.getNickname());
+
+        Member member = memberRepository.save(
+                Member.createMember(
+                        signupRequestDTO.getEmail(),
+                        passwordEncoder.encode(signupRequestDTO.getPassword()),
+                        signupRequestDTO.getNickname(),
+                        signupRequestDTO.getImage()
+                )
+        );
+
+        memberRoleRepository.save(
+                MemberRole.builder()
+                        .memberCode(member.getCode())
+                        .roleCode(RoleType.USER.getOrder())
+                        .build()
+        );
+    }
+}
