--- springboot-monolithic/src/main/java/springbootmonolithic/domain/member/query/dto/MemberDTO.java
@@ -1,18 +1,0 @@
-package springbootmonolithic.domain.member.query.dto;
-
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.ToString;
-
-import java.util.HashSet;
-
-@Getter
-@ToString
-@NoArgsConstructor
-public class MemberDTO {
-    private int code;
-    private String email;
-    private String password;
-    private String nickname;
-    private HashSet<RoleDTO> roleDTO;
-}
