--- springboot-monolithic/src/main/java/springbootmonolithic/domain/member/query/dto/RoleDTO.java
@@ -1,10 +1,0 @@
-package springbootmonolithic.domain.member.query.dto;
-
-import lombok.*;
-
-@Getter
-@Setter
-@NoArgsConstructor
-public class RoleDTO {
-    private String role;
-}
