--- springboot-monolithic/src/main/java/springbootmonolithic/domain/member/query/mapper/AuthMapper.java
@@ -1,10 +1,0 @@
-package springbootmonolithic.domain.member.query.mapper;
-
-import org.apache.ibatis.annotations.Mapper;
-import springbootmonolithic.domain.member.query.dto.MemberDTO;
-
-@Mapper
-public interface AuthMapper {
-    boolean isExistsByEmail(String duplicatedEmail);
-    MemberDTO selectMemberByEmailWithAuthorities(String email);
-}
