+++ springboot-monolithic/src/main/java/springbootmonolithic/domain/member/query/service/MemberQueryService.java
@@ -1,5 +1,7 @@
 package springbootmonolithic.domain.member.query.service;
 
 public interface MemberQueryService {
+    void validateEmail(String email);
+    boolean checkEmailDuplicated(String email);
     void checkNicknameDuplicated(String nickname);
 }
