+++ springboot-monolithic/src/main/java/springbootmonolithic/domain/member/query/service/MemberQueryServiceImpl.java
@@ -4,6 +4,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import springbootmonolithic.domain.member.query.mapper.MemberMapper;
+import springbootmonolithic.exception.EmailDuplicatedException;
 import springbootmonolithic.exception.NicknameDuplicatedException;
 
 @Service
@@ -13,6 +14,20 @@ public class MemberQueryServiceImpl implements MemberQueryService {
     @Autowired
     public MemberQueryServiceImpl(MemberMapper memberMapper) {
         this.memberMapper = memberMapper;
+    }
+
+    @Override
+    @Transactional
+    public void validateEmail(String email) {
+        if (checkEmailDuplicated(email)) {
+            throw new EmailDuplicatedException("이미 존재하는 이메일입니다.");
+        }
+    }
+
+    @Override
+    @Transactional
+    public boolean checkEmailDuplicated(String email) {
+        return memberMapper.isExistsByEmail(email);
     }
 
     @Override
