+++ springboot-monolithic/src/main/java/springbootmonolithic/security/domain/dto/RoleDTO.java
@@ -1,0 +1,10 @@
+package springbootmonolithic.security.domain.dto;
+
+import lombok.*;
+
+@Getter
+@Setter
+@NoArgsConstructor
+public class RoleDTO {
+    private String role;
+}
