+++ springboot-monolithic/src/main/java/springbootmonolithic/security/domain/mapper/AuthMapper.java
@@ -1,0 +1,9 @@
+package springbootmonolithic.security.domain.mapper;
+
+import org.apache.ibatis.annotations.Mapper;
+import springbootmonolithic.security.domain.dto.MemberDTO;
+
+@Mapper
+public interface AuthMapper {
+    MemberDTO selectMemberByEmailWithAuthorities(String email);
+}
