+++ springboot-monolithic/src/main/java/springbootmonolithic/security/domain/service/AuthQueryService.java
@@ -1,0 +1,9 @@
+package springbootmonolithic.security.domain.service;
+
+import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.security.core.userdetails.UserDetailsService;
+
+public interface AuthQueryService extends UserDetailsService {
+    @Override
+    UserDetails loadUserByUsername(String email);
+}
