--- springboot-monolithic/src/main/resources/springbootmonolithic/domain/member/query/mapper/AuthMapper.xml
@@ -1,37 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" ?>
-<!DOCTYPE mapper
-        PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
-        "https://mybatis.org/dtd/mybatis-3-mapper.dtd">
-<mapper namespace="springbootmonolithic.domain.member.query.mapper.AuthMapper">
-    <resultMap id="memberWithRolesMap" type="springbootmonolithic.domain.member.query.dto.MemberDTO">
-        <id property="code" column="CODE"/>
-        <result property="email" column="EMAIL"/>
-        <result property="password" column="PASSWORD"/>
-        <result property="nickname" column="NICKNAME"/>
-        <collection property="roleDTO" ofType="springbootmonolithic.domain.member.query.dto.RoleDTO">
-            <result property="role" column="TYPE"/>
-        </collection>
-    </resultMap>
-    <select id="isExistsByEmail" resultType="_boolean" parameterType="string">
-        SELECT
-               EXISTS (
-                SELECT 1
-                FROM member M
-                WHERE M.email = #{ email }
-        )
-    </select>
-    <select id="selectMemberByEmailWithAuthorities" resultMap="memberWithRolesMap" parameterType="string">
-        SELECT
-               M.CODE
-             , M.EMAIL
-             , M.PASSWORD
-             , M.NICKNAME
-             , R.TYPE
-          FROM member M
-         INNER JOIN member_role MR
-            ON (M.CODE = MR.MEMBER_CODE)
-         INNER JOIN role R
-            ON (MR.ROLE_CODE = R.CODE)
-         WHERE M.EMAIL = #{ email }
-    </select>
-</mapper>
