+++ springboot-monolithic/src/main/resources/springbootmonolithic/security/domain/mapper/AuthMapper.xml
@@ -1,0 +1,29 @@
+<?xml version="1.0" encoding="UTF-8" ?>
+<!DOCTYPE mapper
+        PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
+        "https://mybatis.org/dtd/mybatis-3-mapper.dtd">
+<mapper namespace="springbootmonolithic.security.domain.mapper.AuthMapper">
+    <resultMap id="memberWithRolesMap" type="springbootmonolithic.security.domain.dto.MemberDTO">
+        <id property="code" column="CODE"/>
+        <result property="email" column="EMAIL"/>
+        <result property="password" column="PASSWORD"/>
+        <result property="nickname" column="NICKNAME"/>
+        <collection property="roleDTO" ofType="springbootmonolithic.security.domain.dto.RoleDTO">
+            <result property="role" column="TYPE"/>
+        </collection>
+    </resultMap>
+    <select id="selectMemberByEmailWithAuthorities" resultMap="memberWithRolesMap" parameterType="string">
+        SELECT
+               M.CODE
+             , M.EMAIL
+             , M.PASSWORD
+             , M.NICKNAME
+             , R.TYPE
+          FROM member M
+         INNER JOIN member_role MR
+            ON (M.CODE = MR.MEMBER_CODE)
+         INNER JOIN role R
+            ON (MR.ROLE_CODE = R.CODE)
+         WHERE M.EMAIL = #{ email }
+    </select>
+</mapper>
