+++ springboot-monolithic/src/test/java/springbootmonolithic/integration/domain/member/command/controller/MemberControllerTests.java
@@ -1,0 +1,51 @@
+package springbootmonolithic.integration.domain.member.command.controller;
+
+import com.fasterxml.jackson.databind.ObjectMapper;
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.http.MediaType;
+import org.springframework.test.web.servlet.MockMvc;
+import org.springframework.transaction.annotation.Transactional;
+import springbootmonolithic.domain.member.command.application.dto.request.SignUpRequestDTO;
+
+import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
+import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.print;
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.*;
+
+@SpringBootTest
+@AutoConfigureMockMvc
+@Transactional
+class MemberControllerTests {
+
+    @Autowired
+    private MockMvc mockMvc;
+
+    @Autowired
+    private ObjectMapper objectMapper;
+
+    @Test
+    @DisplayName("회원가입 - 성공")
+    void shouldReturnOkWhenSignUpSuccessful() throws Exception {
+        SignUpRequestDTO signUpRequestDTO = new SignUpRequestDTO(
+                "test@example.com",
+                "password123",
+                "testNickname",
+                "profile-image-url"
+        );
+
+        mockMvc.perform(
+                post("/api/v1/auth/signup")
+                        .contentType(MediaType.APPLICATION_JSON)
+                        .content(objectMapper.writeValueAsString(signUpRequestDTO))
+        )
+                .andExpect(status().isOk())
+                .andExpect(content().contentTypeCompatibleWith(MediaType.APPLICATION_JSON))
+                .andExpect(jsonPath("$.message").value("회원가입 성공"))
+                .andExpect(jsonPath("$.result").value(true))
+                .andExpect(jsonPath("$.timestamp").isNotEmpty())
+                .andDo(print());
+    }
+}
